refactor: 페이지 컴포넌트 폴더명 변경
