+++ src/pages/SectionManagement/SectionManagementPage.styles.ts
